[not verified] Move test utils to a better location
